Consolidate compilation and sharding scripts into compile.py and shard_weights.py
Address reviewer feedback: combine multiple compilation scripts
(compile_transformer.py, compile_transformer_halfres.py, compile_gemma3.py,
compile_vae_23.py) into unified compile.py with subcommands (transformer,
encoder, vae). Combine shard_backbone_weights.py and shard_gemma3_weights.py
into shard_weights.py with subcommands (backbone, encoder).

7 scripts reduced to 3: compile.py, shard_weights.py, compile_benchmark.py.
Update README with new commands and SDK 2.29 DLAMI references.
